More adjustments and an emergency stop
More adjustments and an emergency stop. Added an emergency stop function whose purpose is to stop both motors when the connection to the weather station or the top controller is lost, or when the top controller is not sending data or the data is faulty (SensorData.FO_bad_signal_fault). When it hapens faultFlag = 4